+++ integration_tests/regtest.rs
@@ -11,7 +11,7 @@ use temp_dir::TempDir;
 use tokio::time::sleep;
 
 use bip300301_enforcer_lib::{
-    bins,
+    bins::{self, CommandExt as _},
     proto::{
         self,
         common::{ConsensusHex, Hex, ReverseHex},
@@ -28,7 +28,7 @@ use bip300301_enforcer_lib::{
 };
 use tokio_stream::wrappers::IntervalStream;
 
-use crate::util::{self, drop_temp_dir, AsyncTrial, BinPaths, CommandExt as _};
+use crate::util::{self, AsyncTrial, BinPaths};
 
 #[derive(Clone, Copy, Debug)]
 enum MiningMode {
@@ -118,7 +118,7 @@ async fn setup(bin_paths: &BinPaths, enable_mempool: bool) -> anyhow::Result<Pos
     std::fs::create_dir(&enforcer_dir)?;
     tracing::info!("Enforcer dir: {}", enforcer_dir.display());
     tracing::debug!("Starting bitcoin node");
-    let bitcoind = bins::Bitcoind {
+    let bitcoind = util::Bitcoind {
         path: bin_paths.bitcoind.clone(),
         data_dir: bitcoin_dir,
         listen_port: reserved_ports.bitcoind_listen.port(),
@@ -922,7 +922,7 @@ async fn test(bin_paths: &BinPaths, mode: Mode) -> anyhow::Result<()> {
     let () = withdraw_succeed(&mut post_setup, mode.mining_mode()).await?;
     tracing::info!("Withdrawal succeeded");
     tracing::info!("Removing {}", post_setup.out_dir.path().display());
-    drop_temp_dir(&post_setup.out_dir)?;
+    post_setup.out_dir.cleanup()?;
     drop(post_setup);
     Ok(())
 }
